+++ src/main/php/lang/reflect/Parameter.class.php
@@ -98,6 +98,13 @@ class Parameter {
     }
 
     if ($t= $this->_reflect->getType()) {
+      if ($t instanceof \ReflectionUnionType) {
+        $union= '';
+        foreach ($t->getTypes() as $u) {
+          $union.= '|'.$u->getName();
+        }
+        return substr($union, 1);
+      }
       return PHP_VERSION_ID >= 70100 ? $t->getName() : $t->__toString();
     } else {
       return 'var';